# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "오픈마켓 API",
    "description": "오픈마켓 API Server입니다.<br>\n      <a href=\"/\">버전별 변경사항 확인</a>"
  },
  "servers": [
    {
      "url": "https://api.fesp.shop",
      "description": ""
    },
    {
      "url": "http://localhost",
      "description": "로컬 테스트"
    }
  ],
...
  "paths": {
...
    "/orders/": {
      "post": {
        "tags": [
          "구매"
        ],
        "summary": "상품 구매",
        "description": "상품을 구매합니다.",
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orderCreateRes"
                },
                "examples": {
                  "필수 속성만 지정한 경우": {
                    "$ref": "#/components/examples/createOrderRes"
                  },
                  "추가 속성이 지정된 경우": {
                    "$ref": "#/components/examples/createOrderWithExtraRes"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "422": {
            "description": "입력값 검증 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error422"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "구매 정보가 저장된 객체입니다.<br>products 속성은 필수이며 객체 배열입니다.<br>배열의 요소인 객체는 다음과 같은 필수 정보를 포함해야 하고 추가 속성은 자유롭게 지정하면 됩니다.<br>_id: 상품 id<br>quantity: 구매 수량",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/orderCreate"
              },
              "examples": {
                "필수 속성만 지정한 경우": {
                  "$ref": "#/components/examples/createOrder"
                },
                "추가 속성이 지정된 경우": {
                  "$ref": "#/components/examples/createOrderWithExtra"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "구매"
        ],
        "summary": "구매 목록 조회",
        "description": "자신의 구매 목록을 조회합니다.",
        "parameters": [
          {
            "name": "keyword",
            "description": "상품명 검색어",
            "in": "query",
            "example": "레고",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "custom",
            "description": "custom 검색 조건",
            "in": "query",
            "example": "{'cost.total':{'$gte':40000}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "description": "페이지",
            "in": "query",
            "example": 2,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "description": "한 페이지당 항목 수",
            "in": "query",
            "example": 10,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "description": "정렬(내림차순: -1, 오름차순: 1)",
            "in": "query",
            "example": "{'cost.total': -1}",
            "default": "{'createdAt': -1}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orderListRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
    "/orders/{_id}": {
      "get": {
        "tags": [
          "구매"
        ],
        "summary": "구매 상세 조회",
        "description": "구매 상세 내역을 조회합니다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "구매 id",
            "example": 4
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orderInfoRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "본인의 구매 id가 아니거나 존재하지 않는 구매 id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      },
      "patch": {
        "tags": [
          "구매"
        ],
        "summary": "주문별 주문 상태 수정",
        "description": "주문별로 주문 상태를 수정합니다.(반품 요청, 환불 요청 등)<br>하나의 주문에 하나의 상품만 있을 경우 주문별로 주문 상태를 관리하고 싶을 때 사용합니다.",
        "parameters": [
          {
            "name": "_id",
            "description": "구매 id",
            "in": "path",
            "example": 2,
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/updateOrderRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "본인의 구매 id가 아니거나 존재하지 않는 구매 id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "주문 상태 정보",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/updateOrder"
              }
            }
          }
        }
      }
    },
    "/orders/{_id}/products/{product_id}": {
      "patch": {
        "tags": [
          "구매"
        ],
        "summary": "상품별 주문 상태 수정",
        "description": "상품별로 주문 상태를 수정합니다.(반품 요청, 환불 요청 등)<br>여러 판매자의 상품을 한번에 주문하고 결제했을 경우 하나의 주문에 여러 상품이 있고 각 상품별로 주문 상태를 관리하고 싶을 때 사용합니다.",
        "parameters": [
          {
            "name": "_id",
            "description": "구매 id",
            "in": "path",
            "example": 2,
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "product_id",
            "description": "상품 id",
            "in": "path",
            "example": 3,
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/updateOrderProductRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "본인의 구매 id가 아니거나 존재하지 않는 구매 id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "주문 상태 정보",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/updateOrderProduct"
              }
            }
          }
        }
      }
    },
...
    "/replies/seller/{seller_id}": {
      "get": {
        "tags": [
          "구매 후기"
        ],
        "summary": "판매자 구매 후기 목록",
        "description": "판매자의 상품별로 등록된 모든 구매 후기 목록을 조회합니다.",
        "parameters": [
          {
            "name": "seller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "판매자 id",
            "example": 2
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sellerReplyListRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        }
      }
    },
...
    "/posts/seller/{_id}": {
      "get": {
        "tags": [
          "게시판"
        ],
        "summary": "판매자의 상품에 등록된 게시글 목록",
        "description": "판매자의 상품에 등록된 게시글 목록을 조회합니다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "판매자 id",
            "example": 2
          },
          {
            "name": "product_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "product_id",
            "description": "상품 id",
            "in": "path",
            "example": 1,
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "description": "게시판 종류",
            "in": "query",
            "default": "post",
            "example": "qna",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "description": "검색어<br>제목과 내용 검색에 사용되는 키워드",
            "in": "query",
            "example": "배송",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "custom",
            "description": "custom 검색 조건",
            "in": "query",
            "example": "{'createdAt': {'$gte': '2024.04', '$lt': '2024.05'}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "description": "페이지",
            "in": "query",
            "example": 2,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "description": "한 페이지당 항목 수",
            "in": "query",
            "example": 10,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "description": "정렬(내림차순: -1, 오름차순: 1)",
            "in": "query",
            "example": "{'createdAt': 1}",
            "default": "{'createdAt': -1}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/postListRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        }
      }
    },
...
    "/seller/orders/": {
      "get": {
        "tags": [
          "주문 관리"
        ],
        "summary": "주문 목록 조회",
        "description": "나에게 주문한 내역을 조회합니다.",
        "parameters": [
          {
            "name": "user_id",
            "description": "주문한 회원 id",
            "in": "query",
            "example": 4,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "state",
            "description": "주문 상태",
            "in": "query",
            "example": "OS020",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "custom",
            "description": "custom 검색 조건",
            "in": "query",
            "example": "{'createdAt': {'$gte': '2024.04', '$lt': '2024.05'}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "description": "페이지",
            "in": "query",
            "example": 2,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "description": "한 페이지당 항목 수",
            "in": "query",
            "example": 10,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "description": "정렬(내림차순: -1, 오름차순: 1)",
            "in": "query",
            "example": "{'createdAt': 1}",
            "default": "{'createdAt': -1}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orderListRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
    "/seller/orders/{_id}": {
      "get": {
        "tags": [
          "주문 관리"
        ],
        "summary": "주문 상세 조회",
        "description": "주문 상세 내역을 조회합니다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "주문 id",
            "example": "2"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orderInfoSellerRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "리소스가 존재하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      },
      "patch": {
        "tags": [
          "주문 관리"
        ],
        "summary": "주문별 주문 상태 수정",
        "description": "주문별로 주문 상태를 수정합니다.(배송 시작, 환불 완료 등)<br>하나의 주문에 하나의 상품만 있을 경우 주문별로 주문 상태를 관리하고 싶을 때 사용합니다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "주문 id",
            "example": 2
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/updateOrderSellerRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "리소스가 존재하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "수정할 주문 정보가 저장된 객체입니다.<br>state: 주문 정보의 state 값으로 지정됩니다.<br>나머지 속성은 주문 정보의 history 속성에 추가됩니다.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateOrderSellerBody"
              }
            }
          }
        }
      }
    },
    "/seller/orders/{_id}/products/{product_id}": {
      "patch": {
        "tags": [
          "주문 관리"
        ],
        "summary": "상품별 주문 상태 수정",
        "description": "상품별로 주문 상태를 수정합니다.(배송 시작, 환불 완료 등)<br>여러 판매자의 상품을 한번에 주문하고 결제했을 경우 하나의 주문에 여러 상품이 있고 각 상품별로 주문 상태를 관리하고 싶을 때 사용합니다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "주문 id",
            "example": 2
          },
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "상품 id",
            "example": 3
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/updateOrderProductSellerRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "리소스가 존재하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "수정할 주문 정보가 저장된 객체입니다.<br>state: 주문한 상품의 state 값으로 지정됩니다.<br>나머지 속성은 주문한 상품의 history 속성에 추가됩니다.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateOrderProductSellerBody"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Access Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      },
      "Refresh Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "error401": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "{인증 실패 사유}"
          },
          "errorName": {
            "type": "string",
            "example": "EmptyAuthorization | TokenExpiredError | JsonWebTokenError"
          }
        }
      },
...
      "orderCreateRes": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 1
          },
          "item": {
            "type": "object",
            "properties": {
              "products": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "number",
                      "example": 4
                    },
                    "quantity": {
                      "type": "number",
                      "example": 2
                    },
                    "name": {
                      "type": "string",
                      "example": "레고 테크닉 42151 부가티 볼리드"
                    },
                    "image": {
                      "type": "string",
                      "example": "/uploads/sample-bugatti.png"
                    },
                    "price": {
                      "type": "number",
                      "example": 90000
                    }
                  }
                }
              },
              "address": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "학교"
                  },
                  "value": {
                    "type": "string",
                    "example": "서울시 강남구 역삼동 234"
                  }
                }
              },
              "user_id": {
                "type": "number",
                "example": 2
              },
              "_id": {
                "type": "number",
                "example": 4
              },
              "createdAt": {
                "type": "string",
                "example": "2023.11.22 08:41:28"
              },
              "cost": {
                "type": "object",
                "properties": {
                  "products": {
                    "type": "number",
                    "example": 205000
                  },
                  "shippingFees": {
                    "type": "number",
                    "example": 6000
                  },
                  "total": {
                    "type": "number",
                    "example": 211000
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  }
}